Renamed file_helper.js to tempdir.js
